Do the CA-OCSP-CRL shuffle
