+++ ocsp/ocsp_test.go
@@ -1,0 +1,487 @@
+package ocsp
+
+import (
+	"context"
+	"crypto"
+	"crypto/x509"
+	"encoding/asn1"
+	"encoding/hex"
+	"fmt"
+	"os"
+	"testing"
+	"time"
+
+	"github.com/jmhodges/clock"
+	"github.com/letsencrypt/boulder/ca"
+	capb "github.com/letsencrypt/boulder/ca/proto"
+	"github.com/letsencrypt/boulder/config"
+	"github.com/letsencrypt/boulder/core"
+	corepb "github.com/letsencrypt/boulder/core/proto"
+	"github.com/letsencrypt/boulder/crl"
+	berrors "github.com/letsencrypt/boulder/errors"
+	"github.com/letsencrypt/boulder/features"
+	"github.com/letsencrypt/boulder/goodkey"
+	"github.com/letsencrypt/boulder/issuance"
+	"github.com/letsencrypt/boulder/linter"
+	blog "github.com/letsencrypt/boulder/log"
+	"github.com/letsencrypt/boulder/metrics"
+	"github.com/letsencrypt/boulder/policy"
+	sapb "github.com/letsencrypt/boulder/sa/proto"
+	"github.com/letsencrypt/boulder/test"
+	"github.com/prometheus/client_golang/prometheus"
+	"golang.org/x/crypto/ocsp"
+	"google.golang.org/grpc"
+	"google.golang.org/protobuf/types/known/emptypb"
+)
+
+var (
+	// * Random public key
+	// * CN = not-example.com
+	// * DNSNames = not-example.com, www.not-example.com
+	CNandSANCSR = mustRead("./testdata/cn_and_san.der.csr")
+
+	// CSR generated by Go:
+	// * Random public key
+	// * CN = not-example.com
+	// * Includes an extensionRequest attribute for a well-formed TLS Feature extension
+	MustStapleCSR = mustRead("./testdata/must_staple.der.csr")
+
+	// CSR generated by Go:
+	// * Random public key
+	// * CN = not-example.com
+	// * Includes an extensionRequest attribute for an unknown extension with an
+	//   empty value. That extension's OID, 2.25.123456789, is on the UUID arc.
+	//   It isn't a real randomly-generated UUID because Go represents the
+	//   components of the OID as 32-bit integers, which aren't large enough to
+	//   hold a real 128-bit UUID; this doesn't matter as far as what we're
+	//   testing here is concerned.
+	UnsupportedExtensionCSR = mustRead("./testdata/unsupported_extension.der.csr")
+
+	// CSR generated by Go:
+	// * Random public key
+	// * CN = not-example.com
+	// * Includes an extensionRequest attribute for the CT poison extension
+	//   with a valid NULL value.
+	CTPoisonExtensionCSR = mustRead("./testdata/ct_poison_extension.der.csr")
+
+	// CSR generated by Go:
+	// * Random public key
+	// * CN = not-example.com
+	// * Includes an extensionRequest attribute for the CT poison extension
+	//   with an invalid empty value.
+	CTPoisonExtensionEmptyCSR = mustRead("./testdata/ct_poison_extension_empty.der.csr")
+
+	// CSR generated by Go:
+	// * Random ECDSA public key.
+	// * CN = [none]
+	// * DNSNames = example.com, example2.com
+	ECDSACSR = mustRead("./testdata/ecdsa.der.csr")
+
+	// OIDExtensionCTPoison is defined in RFC 6962 s3.1.
+	OIDExtensionCTPoison = asn1.ObjectIdentifier{1, 3, 6, 1, 4, 1, 11129, 2, 4, 3}
+
+	// OIDExtensionSCTList is defined in RFC 6962 s3.3.
+	OIDExtensionSCTList = asn1.ObjectIdentifier{1, 3, 6, 1, 4, 1, 11129, 2, 4, 2}
+)
+
+const arbitraryRegID int64 = 1001
+const yamlLoadErrMsg = "Error loading YAML bytes for ECDSA allow list:"
+
+func mustRead(path string) []byte {
+	b, err := os.ReadFile(path)
+	if err != nil {
+		panic(fmt.Sprintf("unable to read %#v: %s", path, err))
+	}
+	return b
+}
+
+type testCtx struct {
+	pa             core.PolicyAuthority
+	ocsp           *OcspImpl
+	crl            *crl.CrlImpl
+	certExpiry     time.Duration
+	certBackdate   time.Duration
+	serialPrefix   int
+	maxNames       int
+	boulderIssuers []*issuance.Issuer
+	keyPolicy      goodkey.KeyPolicy
+	fc             clock.FakeClock
+	stats          prometheus.Registerer
+	signatureCount *prometheus.CounterVec
+	signErrorCount *prometheus.CounterVec
+	logger         *blog.Mock
+}
+
+type mockSA struct {
+	certificate core.Certificate
+}
+
+func (m *mockSA) AddCertificate(ctx context.Context, req *sapb.AddCertificateRequest, _ ...grpc.CallOption) (*emptypb.Empty, error) {
+	m.certificate.DER = req.Der
+	return nil, nil
+}
+
+func (m *mockSA) AddPrecertificate(ctx context.Context, req *sapb.AddCertificateRequest, _ ...grpc.CallOption) (*emptypb.Empty, error) {
+	return &emptypb.Empty{}, nil
+}
+
+func (m *mockSA) AddSerial(ctx context.Context, req *sapb.AddSerialRequest, _ ...grpc.CallOption) (*emptypb.Empty, error) {
+	return &emptypb.Empty{}, nil
+}
+
+func (m *mockSA) GetCertificate(ctx context.Context, req *sapb.Serial, _ ...grpc.CallOption) (*corepb.Certificate, error) {
+	return nil, berrors.NotFoundError("cannot find the cert")
+}
+
+func (m *mockSA) GetPrecertificate(ctx context.Context, req *sapb.Serial, _ ...grpc.CallOption) (*corepb.Certificate, error) {
+	return nil, berrors.NotFoundError("cannot find the precert")
+}
+
+// Useful key and certificate files.
+const caKeyFile = "../test/test-ca.key"
+const caCertFile = "../test/test-ca.pem"
+const caCertFile2 = "../test/test-ca2.pem"
+
+var caKey crypto.Signer
+var caCert *issuance.Certificate
+var caCert2 *issuance.Certificate
+var caLinter *linter.Linter
+var caLinter2 *linter.Linter
+var ctx = context.Background()
+
+func init() {
+	var err error
+	caCert, caKey, err = issuance.LoadIssuer(issuance.IssuerLoc{
+		File:     caKeyFile,
+		CertFile: caCertFile,
+	})
+	if err != nil {
+		panic(fmt.Sprintf("Unable to load %q and %q: %s", caKeyFile, caCertFile, err))
+	}
+	caCert2, err = issuance.LoadCertificate(caCertFile2)
+	if err != nil {
+		panic(fmt.Sprintf("Unable to parse %q: %s", caCertFile2, err))
+	}
+	caLinter, _ = linter.New(caCert.Certificate, caKey, []string{"n_subject_common_name_included"})
+	caLinter2, _ = linter.New(caCert2.Certificate, caKey, []string{"n_subject_common_name_included"})
+}
+
+func setup(t *testing.T) *testCtx {
+	features.Reset()
+	fc := clock.NewFake()
+	fc.Add(1 * time.Hour)
+
+	pa, err := policy.New(nil, blog.NewMock())
+	test.AssertNotError(t, err, "Couldn't create PA")
+	err = pa.SetHostnamePolicyFile("../test/hostname-policy.yaml")
+	test.AssertNotError(t, err, "Couldn't set hostname policy")
+
+	boulderProfile := func(rsa, ecdsa bool) *issuance.Profile {
+		res, _ := issuance.NewProfile(
+			issuance.ProfileConfig{
+				AllowMustStaple: true,
+				AllowCTPoison:   true,
+				AllowSCTList:    true,
+				AllowCommonName: true,
+				Policies: []issuance.PolicyInformation{
+					{OID: "2.23.140.1.2.1"},
+				},
+				MaxValidityPeriod:   config.Duration{Duration: time.Hour * 8760},
+				MaxValidityBackdate: config.Duration{Duration: time.Hour},
+			},
+			issuance.IssuerConfig{
+				UseForECDSALeaves: ecdsa,
+				UseForRSALeaves:   rsa,
+				IssuerURL:         "http://not-example.com/issuer-url",
+				OCSPURL:           "http://not-example.com/ocsp",
+				CRLURL:            "http://not-example.com/crl",
+			},
+		)
+		return res
+	}
+	boulderIssuers := []*issuance.Issuer{
+		// Must list ECDSA-only issuer first, so it is the default for ECDSA.
+		{
+			Cert:    caCert2,
+			Signer:  caKey,
+			Profile: boulderProfile(false, true),
+			Linter:  caLinter2,
+			Clk:     fc,
+		},
+		{
+			Cert:    caCert,
+			Signer:  caKey,
+			Profile: boulderProfile(true, true),
+			Linter:  caLinter,
+			Clk:     fc,
+		},
+	}
+
+	keyPolicy := goodkey.KeyPolicy{
+		AllowRSA:           true,
+		AllowECDSANISTP256: true,
+		AllowECDSANISTP384: true,
+	}
+	signatureCount := prometheus.NewCounterVec(
+		prometheus.CounterOpts{
+			Name: "signatures",
+			Help: "Number of signatures",
+		},
+		[]string{"purpose", "issuer"})
+	signErrorCount := prometheus.NewCounterVec(prometheus.CounterOpts{
+		Name: "signature_errors",
+		Help: "A counter of signature errors labelled by error type",
+	}, []string{"type"})
+
+	ocsp, err := NewOCSPImpl(
+		boulderIssuers,
+		time.Hour,
+		0,
+		time.Second,
+		blog.NewMock(),
+		metrics.NoopRegisterer,
+		signatureCount,
+		signErrorCount,
+		fc,
+	)
+	test.AssertNotError(t, err, "Failed to create ocsp impl")
+
+	crl, err := crl.NewCRLImpl(
+		boulderIssuers,
+		time.Hour,
+		"http://c.boulder.test",
+		100,
+		blog.NewMock(),
+	)
+	test.AssertNotError(t, err, "Failed to create crl impl")
+
+	return &testCtx{
+		pa:             pa,
+		ocsp:           ocsp,
+		crl:            crl,
+		certExpiry:     8760 * time.Hour,
+		certBackdate:   time.Hour,
+		serialPrefix:   17,
+		maxNames:       2,
+		boulderIssuers: boulderIssuers,
+		keyPolicy:      keyPolicy,
+		fc:             fc,
+		stats:          metrics.NoopRegisterer,
+		signatureCount: signatureCount,
+		signErrorCount: signErrorCount,
+		logger:         blog.NewMock(),
+	}
+}
+
+func serial(t *testing.T) []byte {
+	serial, err := hex.DecodeString("aabbccddeeffaabbccddeeff000102030405")
+	if err != nil {
+		t.Fatal(err)
+	}
+	return serial
+
+}
+
+func TestOCSP(t *testing.T) {
+	testCtx := setup(t)
+	ca, err := ca.NewCertificateAuthorityImpl(
+		&mockSA{},
+		testCtx.pa,
+		testCtx.ocsp,
+		testCtx.boulderIssuers,
+		nil,
+		testCtx.certExpiry,
+		testCtx.certBackdate,
+		testCtx.serialPrefix,
+		testCtx.maxNames,
+		testCtx.keyPolicy,
+		nil,
+		testCtx.logger,
+		testCtx.stats,
+		testCtx.signatureCount,
+		testCtx.signErrorCount,
+		testCtx.fc)
+	test.AssertNotError(t, err, "Failed to create CA")
+	ocspi := testCtx.ocsp
+
+	// Issue a certificate from the RSA issuer caCert, then check OCSP comes from the same issuer.
+	rsaIssuerID := ca.issuers.byAlg[x509.RSA].ID()
+	rsaCertPB, err := ca.IssuePrecertificate(ctx, &capb.IssueCertificateRequest{Csr: CNandSANCSR, RegistrationID: arbitraryRegID})
+	test.AssertNotError(t, err, "Failed to issue certificate")
+	rsaCert, err := x509.ParseCertificate(rsaCertPB.DER)
+	test.AssertNotError(t, err, "Failed to parse rsaCert")
+	rsaOCSPPB, err := ocspi.GenerateOCSP(ctx, &capb.GenerateOCSPRequest{
+		Serial:   core.SerialToString(rsaCert.SerialNumber),
+		IssuerID: int64(rsaIssuerID),
+		Status:   string(core.OCSPStatusGood),
+	})
+	test.AssertNotError(t, err, "Failed to generate OCSP")
+	rsaOCSP, err := ocsp.ParseResponse(rsaOCSPPB.Response, caCert.Certificate)
+	test.AssertNotError(t, err, "Failed to parse / validate OCSP for rsaCert")
+	test.AssertEquals(t, rsaOCSP.Status, 0)
+	test.AssertEquals(t, rsaOCSP.RevocationReason, 0)
+	test.AssertEquals(t, rsaOCSP.SerialNumber.Cmp(rsaCert.SerialNumber), 0)
+
+	// Issue a certificate from the ECDSA issuer caCert2, then check OCSP comes from the same issuer.
+	ecdsaIssuerID := ca.issuers.byAlg[x509.ECDSA].ID()
+	ecdsaCertPB, err := ca.IssuePrecertificate(ctx, &capb.IssueCertificateRequest{Csr: ECDSACSR, RegistrationID: arbitraryRegID})
+	test.AssertNotError(t, err, "Failed to issue certificate")
+	ecdsaCert, err := x509.ParseCertificate(ecdsaCertPB.DER)
+	test.AssertNotError(t, err, "Failed to parse ecdsaCert")
+	ecdsaOCSPPB, err := ocspi.GenerateOCSP(ctx, &capb.GenerateOCSPRequest{
+		Serial:   core.SerialToString(ecdsaCert.SerialNumber),
+		IssuerID: int64(ecdsaIssuerID),
+		Status:   string(core.OCSPStatusGood),
+	})
+	test.AssertNotError(t, err, "Failed to generate OCSP")
+	ecdsaOCSP, err := ocsp.ParseResponse(ecdsaOCSPPB.Response, caCert2.Certificate)
+	test.AssertNotError(t, err, "Failed to parse / validate OCSP for ecdsaCert")
+	test.AssertEquals(t, ecdsaOCSP.Status, 0)
+	test.AssertEquals(t, ecdsaOCSP.RevocationReason, 0)
+	test.AssertEquals(t, ecdsaOCSP.SerialNumber.Cmp(ecdsaCert.SerialNumber), 0)
+
+	// GenerateOCSP with a bad IssuerID should fail.
+	_, err = ocspi.GenerateOCSP(context.Background(), &capb.GenerateOCSPRequest{
+		Serial:   core.SerialToString(rsaCert.SerialNumber),
+		IssuerID: int64(666),
+		Status:   string(core.OCSPStatusGood),
+	})
+	test.AssertError(t, err, "GenerateOCSP didn't fail with invalid IssuerID")
+
+	// GenerateOCSP with a bad Serial should fail.
+	_, err = ocspi.GenerateOCSP(context.Background(), &capb.GenerateOCSPRequest{
+		Serial:   "BADDECAF",
+		IssuerID: int64(rsaIssuerID),
+		Status:   string(core.OCSPStatusGood),
+	})
+	test.AssertError(t, err, "GenerateOCSP didn't fail with invalid Serial")
+
+	// GenerateOCSP with a valid-but-nonexistent Serial should *not* fail.
+	_, err = ocspi.GenerateOCSP(context.Background(), &capb.GenerateOCSPRequest{
+		Serial:   "03DEADBEEFBADDECAFFADEFACECAFE30",
+		IssuerID: int64(rsaIssuerID),
+		Status:   string(core.OCSPStatusGood),
+	})
+	test.AssertNotError(t, err, "GenerateOCSP failed with fake-but-valid Serial")
+}
+
+// Set up an ocspLogQueue with a very long period and a large maxLen,
+// to ensure any buffered entries get flushed on `.stop()`.
+func TestOcspLogFlushOnExit(t *testing.T) {
+	t.Parallel()
+	log := blog.NewMock()
+	stats := metrics.NoopRegisterer
+	queue := newOCSPLogQueue(4000, 10000*time.Millisecond, stats, log)
+	go queue.loop()
+	queue.enqueue(serial(t), time.Now(), ocsp.Good, ocsp.Unspecified)
+	queue.stop()
+
+	expected := []string{
+		"INFO: [AUDIT] OCSP signed: aabbccddeeffaabbccddeeff000102030405:_,",
+	}
+	test.AssertDeepEquals(t, log.GetAll(), expected)
+}
+
+// Ensure log lines are sent when they exceed maxLen.
+func TestOcspFlushOnLength(t *testing.T) {
+	t.Parallel()
+	log := blog.NewMock()
+	stats := metrics.NoopRegisterer
+	queue := newOCSPLogQueue(100, 100*time.Millisecond, stats, log)
+	go queue.loop()
+	for i := 0; i < 5; i++ {
+		queue.enqueue(serial(t), time.Now(), ocsp.Good, ocsp.Unspecified)
+	}
+	queue.stop()
+
+	expected := []string{
+		"INFO: [AUDIT] OCSP signed: aabbccddeeffaabbccddeeff000102030405:_,aabbccddeeffaabbccddeeff000102030405:_,",
+		"INFO: [AUDIT] OCSP signed: aabbccddeeffaabbccddeeff000102030405:_,aabbccddeeffaabbccddeeff000102030405:_,",
+		"INFO: [AUDIT] OCSP signed: aabbccddeeffaabbccddeeff000102030405:_,",
+	}
+	test.AssertDeepEquals(t, log.GetAll(), expected)
+}
+
+// Ensure log lines are sent after a timeout.
+func TestOcspFlushOnTimeout(t *testing.T) {
+	t.Parallel()
+	log := blog.NewWaitingMock()
+	stats := metrics.NoopRegisterer
+	queue := newOCSPLogQueue(90000, 10*time.Millisecond, stats, log)
+
+	go queue.loop()
+	queue.enqueue(serial(t), time.Now(), ocsp.Good, ocsp.Unspecified)
+
+	expected := "INFO: [AUDIT] OCSP signed: aabbccddeeffaabbccddeeff000102030405:_,"
+	logLines, err := log.WaitForMatch("OCSP signed", 50*time.Millisecond)
+	test.AssertNotError(t, err, "error in mock log")
+	test.AssertDeepEquals(t, logLines, expected)
+	queue.stop()
+}
+
+// If the deadline passes and nothing has been logged, we should not log a blank line.
+func TestOcspNoEmptyLines(t *testing.T) {
+	t.Parallel()
+	log := blog.NewMock()
+	stats := metrics.NoopRegisterer
+	queue := newOCSPLogQueue(90000, 10*time.Millisecond, stats, log)
+
+	go queue.loop()
+	time.Sleep(50 * time.Millisecond)
+	queue.stop()
+
+	test.AssertDeepEquals(t, log.GetAll(), []string{})
+}
+
+// If the maxLogLen is shorter than one entry, log everything immediately.
+func TestOcspLogWhenMaxLogLenIsShort(t *testing.T) {
+	t.Parallel()
+	log := blog.NewMock()
+	stats := metrics.NoopRegisterer
+	queue := newOCSPLogQueue(3, 10000*time.Millisecond, stats, log)
+	go queue.loop()
+	queue.enqueue(serial(t), time.Now(), ocsp.Good, ocsp.Unspecified)
+	queue.stop()
+
+	expected := []string{
+		"INFO: [AUDIT] OCSP signed: aabbccddeeffaabbccddeeff000102030405:_,",
+	}
+	test.AssertDeepEquals(t, log.GetAll(), expected)
+}
+
+// Enqueueing entries after stop causes panic.
+func TestOcspLogPanicsOnEnqueueAfterStop(t *testing.T) {
+	t.Parallel()
+
+	log := blog.NewMock()
+	stats := metrics.NoopRegisterer
+	queue := newOCSPLogQueue(4000, 10000*time.Millisecond, stats, log)
+	go queue.loop()
+	queue.stop()
+
+	defer func() {
+		if r := recover(); r == nil {
+			t.Errorf("The code did not panic")
+		}
+	}()
+
+	queue.enqueue(serial(t), time.Now(), ocsp.Good, ocsp.Unspecified)
+}
+
+// Ensure revoke reason gets set.
+func TestOcspRevokeReasonIsSet(t *testing.T) {
+	t.Parallel()
+	log := blog.NewMock()
+	stats := metrics.NoopRegisterer
+	queue := newOCSPLogQueue(100, 100*time.Millisecond, stats, log)
+	go queue.loop()
+
+	queue.enqueue(serial(t), time.Now(), ocsp.Revoked, ocsp.KeyCompromise)
+	queue.enqueue(serial(t), time.Now(), ocsp.Revoked, ocsp.CACompromise)
+	queue.stop()
+
+	expected := []string{
+		"INFO: [AUDIT] OCSP signed: aabbccddeeffaabbccddeeff000102030405:1,aabbccddeeffaabbccddeeff000102030405:2,",
+	}
+	test.AssertDeepEquals(t, log.GetAll(), expected)
+}
